backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
prophet==1.1.5
ydata-profiling==4.6.0
httpx==0.25.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
psycopg2-binary==2.9.9
